Array: remove ctor with std::initializer_list arg
In cases where the Json::Array ctor is called using uniform
initialization there is an ambiguity when passing only one element.

In these cases the list initialization ctor is called instead of the
probably intended copy ctor.

Replace the list initialization ctor by a factory function for the other cases.